Fix "-target" completion
I've removed the old option_targets() function and instead overwritten
the original statelist(), which is used for getting a list of available
resources.

Unlike Terraform, Terragrunt creates a random cache directory for the
Terraform files. The statelist() function will lookup the name of the
cache dir, jump there and then execute `terraform state list`, which is
then used as the completion source.